Preserve helper entitlements after Electron signing
